+++ Model/Config/Source/CreditCardFormStyle.php
@@ -10,13 +10,17 @@
 namespace SDM\Altapay\Model\Config\Source;
 
 use Magento\Framework\Option\ArrayInterface;
+<<<<<<< HEAD
 use Magento\Store\Model\ScopeInterface;
 use SDM\Altapay\Model\SystemConfig;
+=======
+>>>>>>> d74338b (Refactor code)
 
 class CreditCardFormStyle implements ArrayInterface
 {
 
     /**
+<<<<<<< HEAD
      * @var SystemConfig
      */
     private $systemConfig;
@@ -32,6 +36,8 @@ class CreditCardFormStyle implements ArrayInterface
     }
 
     /**
+=======
+>>>>>>> d74338b (Refactor code)
      * @var array
      */
     private static $designOptions = [
@@ -47,6 +53,7 @@ class CreditCardFormStyle implements ArrayInterface
      */
     public function toOptionArray()
     {
+<<<<<<< HEAD
         $storeScope = ScopeInterface::SCOPE_STORE;
         $storeCode = $this->systemConfig->resolveCurrentStoreCode();
         $login = $this->systemConfig->getApiConfig('api_log_in', $storeScope, $storeCode);
@@ -59,6 +66,10 @@ class CreditCardFormStyle implements ArrayInterface
         
         $output = [];
         foreach ($options as $key => $label) {
+=======
+        $output = [];
+        foreach (self::$designOptions as $key => $label) {
+>>>>>>> d74338b (Refactor code)
             $output[] = ['value' => $key, 'label' => $label];
         }