+++ CHANGELOG.md
@@ -1,6 +1,14 @@
 # Changelog
 All notable changes to this project will be documented in this file.
 
+## [3.7.0]
+### Added
+- Add option to select charged currency for AltaPay payment processing.
+### Fixed
+- Fix: Amount mismatch issue with Apple Pay.
+
+**Note:** This version contains a breaking change. Please ensure that you capture/refund payment before upgrading to this version.
+
 ## [3.6.13]
 ### Added
 - Add a new design option with a modern look for the Credit Card form.